+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.TeamsChannelsPolicy.Tests.ps1
@@ -27,8 +27,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             $Global:PartialExportFileName = 'c:\TestPath'
 
+            Mock -ModuleName M365DSCUtil -CommandName Confirm-M365DSCDependencies -MockWith {
+            }
+
             Mock -CommandName Save-M365DSCPartialExport -MockWith {
             }
+
             Mock -CommandName New-M365DSCConnection -MockWith {
                 return 'Credentials'
             }